+++ app/src/main/java/com/example/android/fishy/network/models/Product.java
@@ -9,7 +9,7 @@ public class Product {
     public Long id;
     public String fish_name;
     public Double price;
-    public Integer stock;
+    public Double stock;
 
     public String getFish_name() {
         return fish_name;
@@ -27,15 +27,15 @@ public class Product {
         this.price = price;
     }
 
-    public Integer getStock() {
+    public Double getStock() {
         return stock;
     }
 
-    public void setStock(Integer stock) {
+    public void setStock(Double stock) {
         this.stock = stock;
     }
 
-    public Product(String fish_name, Double price, Integer stock){
+    public Product(String fish_name, Double price, Double stock){
         this.fish_name=fish_name;
         this.price=price;
         this.stock=stock;
